+++ Targets/LiveAMessage/Sources/Scenes/Login/ViewModel/LoginViewModel.swift
@@ -12,10 +12,6 @@ protocol LoginViewModelProtocol {
     func authenticate()
 }
 
-protocol AutheticationService: NSObject {
-    func authenticate()
-}
-
 class LoginViewModel: LoginViewModelProtocol {
 
     private let service: AutheticationService
@@ -26,35 +22,5 @@ class LoginViewModel: LoginViewModelProtocol {
 
     func authenticate() {
         service.authenticate()
-    }
-}
-
-class SignInWithAppleAuthorization: NSObject, AutheticationService {
-    func authenticate() {
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName, .email]
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.performRequests()
-    }
-}
-
-extension SignInWithAppleAuthorization: ASAuthorizationControllerDelegate {
-
-    func authorizationController(controller: ASAuthorizationController,
-                                 didCompleteWithAuthorization authorization: ASAuthorization) {
-        if let appleIDCredential = authorization.credential as?
-            ASAuthorizationAppleIDCredential {
-            let userIdentifier = appleIDCredential.user
-            let fullName = appleIDCredential.fullName
-            let email = appleIDCredential.email
-            print("User id is \(userIdentifier) \n Full Name is \(String(describing: fullName)) \n Email id is \(String(describing: email))")
-        }
-    }
-
-    func authorizationController(controller: ASAuthorizationController,
-                                 didCompleteWithError error: Error) {
-        print(error)
     }
 }
