+++ Targets/Networking/Sources/Authentication/AuthenticationService.swift
@@ -7,3 +7,7 @@
 //
 
 import Foundation
+
+protocol AutheticationService: NSObject {
+    func authenticate()
+}
