+++ Targets/Networking/Sources/Authentication/SignInWithAppleAuthorization.swift
@@ -1,0 +1,39 @@
+//
+//  SignInWithAppleAuthorization.swift
+//  Networking
+//
+//  Created by Vinicius Mesquita on 24/09/21.
+//  Copyright © 2021 Akee. All rights reserved.
+//
+
+import AuthenticationServices
+
+class SignInWithAppleAuthorization: NSObject, AutheticationService {
+    func authenticate() {
+        let appleIDProvider = ASAuthorizationAppleIDProvider()
+        let request = appleIDProvider.createRequest()
+        request.requestedScopes = [.fullName, .email]
+        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
+        authorizationController.delegate = self
+        authorizationController.performRequests()
+    }
+}
+
+extension SignInWithAppleAuthorization: ASAuthorizationControllerDelegate {
+
+    func authorizationController(controller: ASAuthorizationController,
+                                 didCompleteWithAuthorization authorization: ASAuthorization) {
+        if let appleIDCredential = authorization.credential as?
+            ASAuthorizationAppleIDCredential {
+            let userIdentifier = appleIDCredential.user
+            let fullName = appleIDCredential.fullName
+            let email = appleIDCredential.email
+            print("User id is \(userIdentifier) \n Full Name is \(String(describing: fullName)) \n Email id is \(String(describing: email))")
+        }
+    }
+
+    func authorizationController(controller: ASAuthorizationController,
+                                 didCompleteWithError error: Error) {
+        print(error)
+    }
+}
